Split test from Factory
